split namespace and change names
