+++ src/Blazor.Extensions.Canvas/Canvas2D/Canvas2DEnums.cs
@@ -1,4 +1,4 @@
-namespace Blazor.Extensions.Canvas
+namespace Blazor.Extensions.Canvas.Canvas2D
 {
     public enum TextAlign
     {
