+++ src/Blazor.Extensions.Canvas/CanvasContextExtensions.cs
@@ -1,10 +1,13 @@
+using Blazor.Extensions.Canvas.Canvas2D;
+using Blazor.Extensions.Canvas.WebGL;
+
 namespace Blazor.Extensions.Canvas
 {
     public static class CanvasContextExtensions
     {
-        public static Canvas2dContext CreateCanvas2D(this BECanvasComponent canvas)
+        public static Canvas2DContext CreateCanvas2D(this BECanvasComponent canvas)
         {
-            return new Canvas2dContext(canvas);
+            return new Canvas2DContext(canvas);
         }
 
         public static WebGLContext CreateWebGL(this BECanvasComponent canvas)
