+++ src/Blazor.Extensions.Canvas/WebGL/WebGLEnums.cs
@@ -1,4 +1,4 @@
-namespace Blazor.Extensions.Canvas
+namespace Blazor.Extensions.Canvas.WebGL
 {
     public enum BufferBits
     {
